% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anabass-package.R
\docType{package}
\name{anabass-package}
\alias{anabass}
\alias{anabass-package}
\title{anabass: Analyse Basic Structure of Sentences for Better Writing}
\description{
Provide an analysis in basic structure of sentences to improve writings. For example finding connetions between senteces and display them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matutosi/anabass}
  \item \url{https://matutosi.github.io/anabass/}
}

}
\author{
\strong{Maintainer}: Toshikazu Matsumura \email{matutosi@gmail.com} (\href{https://orcid.org/0000-0003-2306-6355}{ORCID})

}
\keyword{internal}
