% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{connect_with}
\alias{connect_with}
\title{Find the First Connection for a Given Sentence
Finds the first connection for a given sentence in a list of sentences.}
\usage{
connect_with(sentences, i)
}
\arguments{
\item{sentences}{A list of sentences (character vectors).}

\item{i}{An index of the target sentence}
}
\value{
A list representing the connection
(with elements \code{sentence_i}, \code{sentence_j}, \code{word_i}, and \code{word_j}),
or only \code{sentence_i} when no connection found.
}
\description{
Find the First Connection for a Given Sentence
Finds the first connection for a given sentence in a list of sentences.
}
\examples{
sentences <- list(
  c("there", "is", "a", "quick", "brown", "fox"),
  c("the", "fox", "chase", "a", "lazy", "dog"),
  c("dog", "jump", "over", "the", "moon")
)
connect_with(sentences, 1)
connect_with(sentences, 2)
connect_with(sentences, 3)

}
