% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neko_sentences}
\alias{neko_sentences}
\title{Dammy data for test use}
\format{
A data frame
}
\usage{
neko_sentences
}
\description{
Dammy data for test use
}
\examples{
data(neko_sentences)
neko_sentences
}
\keyword{datasets}
