% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{search_common_word}
\alias{search_common_word}
\title{Find the First Common Word in Two Sentences}
\usage{
search_common_word(sentence_i, sentence_j)
}
\arguments{
\item{sentence_i, sentence_j}{A character vector of sentence.}
}
\value{
A list with two elements or NULL when no common word was found:
\itemize{
\item \strong{common_i:} The index of the common word in the sentence of i.
\item \strong{common_j:} The index of the common word in the sentence of j.
}
}
\description{
Finds the first common word in two sentences and returns its indices in each sentence.
}
\examples{
sentence1 <- c("the", "quick", "brown", "fox")
sentence2 <- c("the", "lazy", "dog")
search_common_word(sentence1, sentence2)

}
