% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{search_index}
\alias{search_index}
\title{Generate search index for multi sentences}
\usage{
search_index(i)
}
\arguments{
\item{i}{An Integer}
}
\value{
A numeric vector
}
\description{
Generate search index for multi sentences
}
\examples{
search_index(1)
search_index(2)
search_index(3)
search_index(4)

}
