% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sentences}
\alias{sentences}
\title{Dammy data for test use from stringr}
\format{
A list
}
\usage{
sentences
}
\description{
Dammy data for test use from stringr
}
\examples{
data(sentences)
sentences
}
\keyword{datasets}
