% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{split_sentence}
\alias{split_sentence}
\title{Split sentence into list}
\usage{
split_sentence(df, group = "sentence")
}
\arguments{
\item{df}{A dataframe}

\item{group}{A string specifying group}
}
\value{
A list of split dataframe
}
\description{
Split sentence into list
}
\examples{
sentences <- list(
  c("the", "quick", "brown", "fox"),
  c("the", "lazy", "dog"),
  c("jumped", "over", "the", "moon")
)
list2df(sentences) |>
  split_sentence()

}
