% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{update_x_pos}
\alias{update_x_pos}
\title{Update \code{x} Positions Based on a Computed Difference
Updates the \code{x_start} and \code{x_end} positions in a data frame based on a computed difference.}
\usage{
update_x_pos(df, diff)
}
\arguments{
\item{df}{A data frame containing sentence information.}

\item{diff}{A list representing the difference in \code{x} values and
the index of the first sentence.}
}
\value{
The modified data frame with updated \code{x_start} and \code{x_end} positions.
}
\description{
Update \code{x} Positions Based on a Computed Difference
Updates the \code{x_start} and \code{x_end} positions in a data frame based on a computed difference.
}
\examples{
sentences <- list(
  c("there", "is", "a", "quick", "brown", "fox"),
  c("the", "fox", "chase", "a", "lazy", "dog"),
  c("dog", "jump", "over", "the", "moon"),
  c("apple", "apricot", "avocado", "banana")
)
df <-
  sentences |>
  list2df() |>
  add_x_pos() |>
  add_index()

for(i in seq_along(sentences)){
  con <- connect_with(sentences, i)
 # if(is.null(con$sentence_j)) next
  diff <- compute_diff_x(df, con)
  df <- update_x_pos(df, diff)
}
print(df, n = Inf)

df |>
  ggplot2::ggplot(ggplot2::aes(x = x_start, y = sentence, label = word)) +
  ggplot2::geom_text() +
  ggplot2::scale_y_reverse() +
  ggplot2::theme_bw()

}
